boot_control: move error def into a new errors.py
