+++ otaclient/app/boot_control/_common.py
@@ -11,13 +11,14 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
-
 r"""Shared utils for boot_controller."""
-from enum import auto, Enum
+
+
 from pathlib import Path
 from subprocess import CalledProcessError
-from typing import Callable, List, Optional, Union
+from typing import List, Optional, Union
+
+from ._errors import BootControlError, MountError, MkfsError, MountFailedReason
 
 from .. import log_util
 from ..configs import config as cfg
@@ -33,86 +34,6 @@ from ..proto import wrapper
 logger = log_util.get_logger(
     __name__, cfg.LOG_LEVEL_TABLE.get(__name__, cfg.DEFAULT_LOG_LEVEL)
 )
-
-
-class _BootControlError(Exception):
-    """Internal used exception type related to boot control errors.
-    NOTE: this exception should not be directly raised to upper caller,
-    it should always be wrapped by a specific OTA Error type.
-    check app.errors for details.
-    """
-
-
-class MountError(_BootControlError):
-    def __init__(
-        self, *args: object, fail_reason: Optional["MountFailedReason"] = None
-    ) -> None:
-        super().__init__(*args)
-        self.fail_reason = fail_reason
-
-
-class ABPartitionError(_BootControlError):
-    ...
-
-
-class MkfsError(_BootControlError):
-    ...
-
-
-class MountFailedReason(Enum):
-    # error code
-    SUCCESS = 0
-    PERMISSIONS_ERROR = 1
-    SYSTEM_ERROR = 2
-    INTERNAL_ERROR = 4
-    USER_INTERRUPT = 8
-    GENERIC_MOUNT_FAILURE = 32
-
-    # custom error code
-    # specific reason for generic mount failure
-    TARGET_NOT_FOUND = auto()
-    TARGET_ALREADY_MOUNTED = auto()
-    MOUNT_POINT_NOT_FOUND = auto()
-    BIND_MOUNT_ON_NON_DIR = auto()
-
-    @classmethod
-    def parse_failed_reason(cls, func: Callable):
-        from functools import wraps
-
-        @wraps(func)
-        def _wrapper(*args, **kwargs):
-            try:
-                return func(*args, **kwargs)
-            except CalledProcessError as e:
-                _called_process_error_str = (
-                    f"{e.cmd=} failed: {e.returncode=}, {e.stderr=}, {e.stdout=}"
-                )
-
-                _reason = cls(e.returncode)
-                if _reason != cls.GENERIC_MOUNT_FAILURE:
-                    raise MountError(
-                        _called_process_error_str,
-                        fail_reason=_reason,
-                    ) from None
-
-                # if the return code is 32, determine the detailed reason
-                # of the mount failure
-                _error_msg, _fail_reason = str(e.stderr), cls.GENERIC_MOUNT_FAILURE
-                if _error_msg.find("already mounted") != -1:
-                    _fail_reason = cls.TARGET_ALREADY_MOUNTED
-                elif _error_msg.find("mount point does not exist") != -1:
-                    _fail_reason = cls.MOUNT_POINT_NOT_FOUND
-                elif _error_msg.find("does not exist") != -1:
-                    _fail_reason = cls.TARGET_NOT_FOUND
-                elif _error_msg.find("Not a directory") != -1:
-                    _fail_reason = cls.BIND_MOUNT_ON_NON_DIR
-
-                raise MountError(
-                    _called_process_error_str,
-                    fail_reason=_fail_reason,
-                ) from None
-
-        return _wrapper
 
 
 class CMDHelperFuncs:
@@ -224,7 +145,7 @@ class CMDHelperFuncs:
         if res := cls._findmnt("/ -o SOURCE -n -f -c"):
             return res
         else:
-            raise _BootControlError("failed to detect current rootfs")
+            raise BootControlError("failed to detect current rootfs")
 
     @classmethod
     def get_mount_point_by_dev(cls, dev: str, *, raise_exception=True) -> str:
@@ -336,7 +257,7 @@ class CMDHelperFuncs:
             logger.warning(_failure_msg)
 
             if not ignore_error:
-                raise _BootControlError(_failure_msg) from None
+                raise BootControlError(_failure_msg) from None
 
     @classmethod
     def mkfs_ext4(cls, dev: str):
